+++ Changelog.md
@@ -6,6 +6,8 @@
   `apollo-client` 2.3.3
   [PR #2105](https://github.com/apollographql/react-apollo/pull/2105)
 
+* Created the `ManualQuery` component: [#1989](https://github.com/apollographql/react-apollo/pull/1989)
+
 ### 2.1.4
 
 - Adds `__typename` for queries made with MockProvider and MockLink